feat: refactor React lint rules to consolidate utility functions in react.js
