+++ lib/rules/react/no-nested-component.js
@@ -10,6 +10,7 @@
  * Configuration:
  * - `ignore`: Array of glob patterns to ignore files.
  */
+const { isComponent } = require("../../utils/react");
 
 const rule = {
   meta: {
@@ -45,41 +46,7 @@ const rule = {
       const filename = context.getFilename();
       if (micromatch.isMatch(filename, ignorePatterns)) return {};
     }
-    // Helper: is a function/class a React component?
-    function isComponent(node) {
-      if (
-        node.type === "FunctionDeclaration" ||
-        node.type === "FunctionExpression" ||
-        node.type === "ArrowFunctionExpression"
-      ) {
-        // Heuristic: name starts with uppercase, or returns JSX
-        if (node.id && /^[A-Z]/.test(node.id.name)) return true;
-        if (
-          node.parent &&
-          node.parent.type === "VariableDeclarator" &&
-          /^[A-Z]/.test(node.parent.id.name)
-        )
-          return true;
-        // Returns JSX (JSXElement or JSXFragment)
-        if (node.body && node.body.type === "BlockStatement") {
-          return node.body.body.some(
-            (stmt) =>
-              stmt.type === "ReturnStatement" &&
-              stmt.argument &&
-              (stmt.argument.type === "JSXElement" ||
-                stmt.argument.type === "JSXFragment")
-          );
-        }
-      }
-      if (
-        node.type === "ClassDeclaration" &&
-        node.id &&
-        /^[A-Z]/.test(node.id.name)
-      ) {
-        return true;
-      }
-      return false;
-    }
+
     // Track component nesting
     let componentStack = [];
     return {
