+++ lib/rules/react/no-unnecessary-fragment.js
@@ -9,7 +9,7 @@
 const {
   getMeaningfulChildren,
   isFragmentElement,
-} = require("../../utils/fragment-helpers");
+} = require("../../utils/react");
 
 const rule = {
   meta: {
