+++ lib/utils/react.js
@@ -43,7 +43,53 @@ function isFragmentElement(node) {
   );
 }
 
+// Helper: is a function/class a React component?
+function isComponent(node) {
+  if (
+    node.type === "FunctionDeclaration" ||
+    node.type === "FunctionExpression" ||
+    node.type === "ArrowFunctionExpression"
+  ) {
+    // Heuristic: name starts with uppercase, or returns JSX
+    if (node.id && /^[A-Z]/.test(node.id.name)) return true;
+    if (
+      node.parent &&
+      node.parent.type === "VariableDeclarator" &&
+      /^[A-Z]/.test(node.parent.id.name)
+    )
+      return true;
+    // Returns JSX (JSXElement or JSXFragment)
+    if (node.body && node.body.type === "BlockStatement") {
+      return node.body.body.some(
+        (stmt) =>
+          stmt.type === "ReturnStatement" &&
+          stmt.argument &&
+          (stmt.argument.type === "JSXElement" ||
+            stmt.argument.type === "JSXFragment")
+      );
+    }
+  }
+  if (
+    node.type === "ClassDeclaration" &&
+    node.id &&
+    /^[A-Z]/.test(node.id.name)
+  ) {
+    return true;
+  }
+  return false;
+}
+
+function isComponentName(name) {
+  return name && /^[A-Z]/.test(name);
+}
+function isHookName(name) {
+  return name && /^use[A-Z0-9]/.test(name);
+}
+
 module.exports = {
   getMeaningfulChildren,
   isFragmentElement,
+  isComponent,
+  isComponentName,
+  isHookName,
 };
